tests, libnet: Move cloudinit to its own package
