+++ tests/BUILD.bazel
@@ -31,7 +31,7 @@ go_library(
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libdv:go_default_library",
-        "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libpod:go_default_library",
         "//tests/libstorage:go_default_library",
@@ -210,6 +210,7 @@ go_test(
         "//tests/libinstancetype:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libpod:go_default_library",
         "//tests/libreplicaset:go_default_library",
