+++ tests/libnet/cloudinit/cloudinit.go
@@ -17,10 +17,12 @@
  *
  */
 
-package libnet
+package cloudinit
 
 import (
 	"fmt"
+
+	"kubevirt.io/kubevirt/tests/libnet"
 
 	"sigs.k8s.io/yaml"
 )
@@ -108,13 +110,13 @@ func WithGateway6(gateway6 string) NetworkDataInterfaceOption {
 
 func WithNameserverFromCluster() NetworkDataInterfaceOption {
 	return func(networkDataInterface *CloudInitInterface) error {
-		dnsServerIP, err := ClusterDNSServiceIP()
+		dnsServerIP, err := libnet.ClusterDNSServiceIP()
 		if err != nil {
 			return fmt.Errorf("failed defining network data nameservers when retrieving cluster DNS service IP: %w", err)
 		}
 		networkDataInterface.Nameservers = CloudInitNameservers{
 			Addresses: []string{dnsServerIP},
-			Search:    SearchDomains(),
+			Search:    libnet.SearchDomains(),
 		}
 		return nil
 	}
