+++ tests/libvmi/BUILD.bazel
@@ -24,7 +24,7 @@ go_library(
         "//tests/containerdisk:go_default_library",
         "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
-        "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit:go_default_library",
         "//tests/testsuite:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
