+++ tests/network/BUILD.bazel
@@ -63,6 +63,7 @@ go_library(
         "//tests/libkvconfig:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
+        "//tests/libnet/cloudinit:go_default_library",
         "//tests/libnet/cluster:go_default_library",
         "//tests/libnet/job:go_default_library",
         "//tests/libnet/service:go_default_library",
