+++ tests/network/vmi_multus.go
@@ -48,6 +48,7 @@ import (
 	"kubevirt.io/kubevirt/tests"
 	"kubevirt.io/kubevirt/tests/console"
 	"kubevirt.io/kubevirt/tests/libnet"
+	"kubevirt.io/kubevirt/tests/libnet/cloudinit"
 	"kubevirt.io/kubevirt/tests/libnode"
 	"kubevirt.io/kubevirt/tests/libvmi"
 	"kubevirt.io/kubevirt/tests/libwait"
@@ -193,10 +194,10 @@ var _ = SIGDescribe("[Serial]Multus", Serial, decorators.Multus, func() {
 			var networkData string
 			BeforeEach(func() {
 				libnet.SkipWhenClusterNotSupportIpv4()
-				networkData, err = libnet.NewNetworkData(
-					libnet.WithEthernet("eth0",
-						libnet.WithDHCP4Enabled(),
-						libnet.WithNameserverFromCluster(),
+				networkData, err = cloudinit.NewNetworkData(
+					cloudinit.WithEthernet("eth0",
+						cloudinit.WithDHCP4Enabled(),
+						cloudinit.WithNameserverFromCluster(),
 					),
 				)
 				Expect(err).NotTo(HaveOccurred())
@@ -280,10 +281,10 @@ var _ = SIGDescribe("[Serial]Multus", Serial, decorators.Multus, func() {
 		Context("VirtualMachineInstance with multus network as default network", func() {
 			It("[test_id:1751]should create a virtual machine with one interface with multus default network definition", func() {
 				libnet.SkipWhenClusterNotSupportIpv4()
-				networkData, err := libnet.NewNetworkData(
-					libnet.WithEthernet("eth0",
-						libnet.WithDHCP4Enabled(),
-						libnet.WithNameserverFromCluster(),
+				networkData, err := cloudinit.NewNetworkData(
+					cloudinit.WithEthernet("eth0",
+						cloudinit.WithDHCP4Enabled(),
+						cloudinit.WithNameserverFromCluster(),
 					),
 				)
 				Expect(err).NotTo(HaveOccurred())
@@ -779,11 +780,11 @@ func checkMacAddress(vmi *v1.VirtualMachineInstance, interfaceName, macAddress s
 }
 
 func cloudInitNetworkDataWithStaticIPsByMac(nicName, macAddress, ipAddress string) string {
-	networkData, err := libnet.NewNetworkData(
-		libnet.WithEthernet(nicName,
-			libnet.WithAddresses(ipAddress),
-			libnet.WithNameserverFromCluster(),
-			libnet.WithMatchingMAC(macAddress),
+	networkData, err := cloudinit.NewNetworkData(
+		cloudinit.WithEthernet(nicName,
+			cloudinit.WithAddresses(ipAddress),
+			cloudinit.WithNameserverFromCluster(),
+			cloudinit.WithMatchingMAC(macAddress),
 		),
 	)
 	ExpectWithOffset(1, err).ToNot(HaveOccurred(), "should successfully create static IPs by mac address cloud init network data")
@@ -791,10 +792,10 @@ func cloudInitNetworkDataWithStaticIPsByMac(nicName, macAddress, ipAddress strin
 }
 
 func cloudInitNetworkDataWithStaticIPsByDevice(deviceName, ipAddress string) string {
-	networkData, err := libnet.NewNetworkData(
-		libnet.WithEthernet(deviceName,
-			libnet.WithAddresses(ipAddress),
-			libnet.WithNameserverFromCluster(),
+	networkData, err := cloudinit.NewNetworkData(
+		cloudinit.WithEthernet(deviceName,
+			cloudinit.WithAddresses(ipAddress),
+			cloudinit.WithNameserverFromCluster(),
 		),
 	)
 	ExpectWithOffset(1, err).ToNot(HaveOccurred(), "should successfully create static IPs by device name cloud init network data")
